Cut down test run time by ~50%
The longest test (hailstone.s) was taking almost triple it should take
(~11s vs. ~4s target run-time).

A single execution of the program from the CLI gives run times around
2200 ms. The test executes it twice (with and without forwarding),
bringing us to a desired test run time of ~4s.

Profiling revealed that a good chunk of the time was spent writing down
the Dinero tracefile, which is used only in few unit tests.

Therefore, this commit makes writing them optional and disables them by
default in unit tests, enabling it only when needed.

The test suite run time went down from ~11s to ~5s (more than acceptable
overhead for the unit test setup and teardown).

+++ src/test/java/org/edumips64/EndToEndTests.java
@@ -153,6 +153,13 @@ public class EndToEndTests extends BaseTest {
    * @param testPath path of the test code.
    */
   private CpuTestStatus runMipsTest(String testPath) throws Exception {
+    return runMipsTest(testPath, false);
+  }
+
+  // Version of runMipsTest that allows to specify whether to write a trace file.
+  // Writing the trace file can unnecessarily slow down the tests, so by default they are not written (see previous
+  // override of this method).
+  private CpuTestStatus runMipsTest(String testPath, boolean writeTracefile) throws Exception {
     log.warning("================================= Starting test " + testPath + " (forwarding: " +
         config.getBoolean(ConfigKey.FORWARDING) + ")");
     cpu.reset();
@@ -160,6 +167,7 @@ public class EndToEndTests extends BaseTest {
     symTab.reset();
     testPath = testsLocation + testPath;
     CycleBuilder builder = new CycleBuilder(cpu);
+    String tracefile = null;
 
     try {
       try {
@@ -183,11 +191,14 @@ public class EndToEndTests extends BaseTest {
       log.warning("================================= Finished test " + testPath);
       log.info(cpu.toString());
 
-      File tmp = File.createTempFile("edumips64", "xdin");
-      tmp.deleteOnExit();
-      LocalWriter w = new LocalWriter(tmp.getAbsolutePath(), false);
-      dinero.writeTraceData(w);
-      w.close();
+      if (writeTracefile) {
+        File tmp = File.createTempFile("edumips64", "xdin");
+        tracefile = tmp.getAbsolutePath();
+        tmp.deleteOnExit();
+        LocalWriter w = new LocalWriter(tmp.getAbsolutePath(), false);
+        dinero.writeTraceData(w);
+        w.close();
+      }
 
       // Check if the transactions in the CycleBuilder are all valid.
       for (CycleElement el : builder.getElementsList()) {
@@ -196,7 +207,7 @@ public class EndToEndTests extends BaseTest {
         }
       }
 
-      return new CpuTestStatus(cpu, tmp.getAbsolutePath());
+      return new CpuTestStatus(cpu, tracefile);
     } finally {
       cpu.reset();
       dinero.reset();
@@ -238,7 +249,7 @@ public class EndToEndTests extends BaseTest {
   }
 
   private void runTestAndCompareTracefileWithGolden(String path) throws Exception {
-    CpuTestStatus s = runMipsTest(path);
+    CpuTestStatus s = runMipsTest(path, true);
     String goldenTrace = testsLocation + path + ".xdin.golden";
 
     String golden = new Scanner(new File(goldenTrace)).useDelimiter("\\A").next();
